refactor: put "enums", "models" and "serializable" into "types" directory
