+++ src/main/java/mrquackduck/imageemojis/commands/EmojisReloadCommand.java
@@ -3,8 +3,8 @@ package mrquackduck.imageemojis.commands;
 import mrquackduck.imageemojis.ImageEmojisPlugin;
 import mrquackduck.imageemojis.configuration.Configuration;
 import mrquackduck.imageemojis.configuration.Permissions;
-import mrquackduck.imageemojis.enums.SuggestionMode;
-import mrquackduck.imageemojis.models.EmojiModel;
+import mrquackduck.imageemojis.types.enums.SuggestionMode;
+import mrquackduck.imageemojis.types.models.EmojiModel;
 import mrquackduck.imageemojis.utils.SuggestionManager;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
