+++ src/main/java/mrquackduck/imageemojis/interfaces/ResourcePackGenerator.java
@@ -1,6 +1,6 @@
 package mrquackduck.imageemojis.interfaces;
 
-import mrquackduck.imageemojis.models.ResourcePack;
+import mrquackduck.imageemojis.types.models.ResourcePack;
 
 public interface ResourcePackGenerator {
     ResourcePack generate();
