+++ src/main/java/mrquackduck/imageemojis/listeners/AnvilRenameListener.java
@@ -3,8 +3,8 @@ package mrquackduck.imageemojis.listeners;
 import mrquackduck.imageemojis.ImageEmojisPlugin;
 import mrquackduck.imageemojis.configuration.Configuration;
 import mrquackduck.imageemojis.configuration.Permissions;
-import mrquackduck.imageemojis.enums.NoPermAction;
-import mrquackduck.imageemojis.models.EmojiModel;
+import mrquackduck.imageemojis.types.enums.NoPermAction;
+import mrquackduck.imageemojis.types.models.EmojiModel;
 import mrquackduck.imageemojis.utils.TextComponentUtil;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.TextComponent;
