+++ src/main/java/mrquackduck/imageemojis/listeners/JoinListener.java
@@ -3,9 +3,9 @@ package mrquackduck.imageemojis.listeners;
 import mrquackduck.imageemojis.ImageEmojisPlugin;
 import mrquackduck.imageemojis.configuration.Configuration;
 import mrquackduck.imageemojis.configuration.Permissions;
-import mrquackduck.imageemojis.enums.EnforcementPolicy;
-import mrquackduck.imageemojis.enums.SuggestionMode;
-import mrquackduck.imageemojis.models.ResourcePack;
+import mrquackduck.imageemojis.types.enums.EnforcementPolicy;
+import mrquackduck.imageemojis.types.enums.SuggestionMode;
+import mrquackduck.imageemojis.types.models.ResourcePack;
 import mrquackduck.imageemojis.utils.SuggestionManager;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
