+++ src/main/java/mrquackduck/imageemojis/listeners/SendCommandListener.java
@@ -3,8 +3,8 @@ package mrquackduck.imageemojis.listeners;
 import mrquackduck.imageemojis.ImageEmojisPlugin;
 import mrquackduck.imageemojis.configuration.Configuration;
 import mrquackduck.imageemojis.configuration.Permissions;
-import mrquackduck.imageemojis.enums.NoPermAction;
-import mrquackduck.imageemojis.models.EmojiModel;
+import mrquackduck.imageemojis.types.enums.NoPermAction;
+import mrquackduck.imageemojis.types.models.EmojiModel;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
