+++ src/main/java/mrquackduck/imageemojis/listeners/SendMessageListener.java
@@ -4,8 +4,8 @@ import io.papermc.paper.event.player.AsyncChatEvent;
 import mrquackduck.imageemojis.ImageEmojisPlugin;
 import mrquackduck.imageemojis.configuration.Configuration;
 import mrquackduck.imageemojis.configuration.Permissions;
-import mrquackduck.imageemojis.enums.NoPermAction;
-import mrquackduck.imageemojis.models.EmojiModel;
+import mrquackduck.imageemojis.types.enums.NoPermAction;
+import mrquackduck.imageemojis.types.models.EmojiModel;
 import mrquackduck.imageemojis.utils.ColorUtil;
 import mrquackduck.imageemojis.utils.TextComponentUtil;
 import net.kyori.adventure.text.Component;
