+++ src/main/java/mrquackduck/imageemojis/types/enums/EnforcementPolicy.java
@@ -1,4 +1,4 @@
-package mrquackduck.imageemojis.enums;
+package mrquackduck.imageemojis.types.enums;
 
 public enum EnforcementPolicy {
     NONE,
